added plugin manager api project moved plugin interface to new project fixed tests
