+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/ConfigurationImpl.java
@@ -27,6 +27,7 @@ import java.util.Set;
 
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.Reference;
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.Service;
+import de.unistuttgart.iaas.amyassist.amy.core.pluginloader.IPlugin;
 import de.unistuttgart.iaas.amyassist.amy.core.pluginloader.PluginLoader;
 
 /**
