+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/configuration/PropertiesProvider.java
@@ -28,10 +28,10 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.Properties;
 
-import de.unistuttgart.iaas.amyassist.amy.core.IPlugin;
 import de.unistuttgart.iaas.amyassist.amy.core.di.Context;
 import de.unistuttgart.iaas.amyassist.amy.core.di.ServiceFactory;
 import de.unistuttgart.iaas.amyassist.amy.core.di.consumer.ServiceFunction;
+import de.unistuttgart.iaas.amyassist.amy.core.pluginloader.IPlugin;
 
 /**
  * Provide the plugin configuration properties as service
