+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/pluginloader/PluginManagerService.java
@@ -33,7 +33,6 @@ import javax.ws.rs.Path;
 
 import org.slf4j.Logger;
 
-import de.unistuttgart.iaas.amyassist.amy.core.IPlugin;
 import de.unistuttgart.iaas.amyassist.amy.core.configuration.ConfigurationLoader;
 import de.unistuttgart.iaas.amyassist.amy.core.di.DependencyInjection;
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.PostConstruct;
