+++ pom.xml
@@ -42,6 +42,8 @@
 		<module>http-server</module>
 		<module>testing-framework</module>
 		<module>core</module>
+		<module>plugin-manager-api</module>
+		<module>rest-resources</module>
 		<module>plugins/alarmclock</module>
 		<module>plugins/example</module>
 		<module>plugins/spotify</module>
