--- rest-resources/.gitignore
@@ -1,1 +1,0 @@
-/target/
